+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 9. Fixed Issue with Cpp2IL Package Cleanup.
 10. Fixed Comparison Issue with Semver.
 11. Fixed Issue with Start Screen Background and Logo not Animating Properly.
+12. Added Start Screen Element Position Customization.
 
 ---